chore: remove the `handleBlockImport` goroutine and introduce `BlockImportHandler` at `digest` pkg
